Add new node flag to track calls for fear analysis
The optimizer may choose to not inline or optimize a call with the expectation
that ValuePropagation will be able to compile-time fold the call to a constant.
We need to track these calls so that, when HCR is enabled, we can generate
OSR guards as appropriate. This is a performance optimization that allows us
to omit HCR guards in some situations.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>