Add facility to split blocks post GRA
After GRA has run, splitting a block becomes more complex because of the
need to correct the GlRegDeps on the blocks. Further, when fixing up the
commoning, it is possible to store values to registers rather than the
using stack temps to improve performance. This commit adds a new postGRA
block split API on OMBlock to facilitate this splitting and use of
registers to hold values.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>